修改接口

See merge request kl-mos-server/hospital!162

+++ kl-environment/src/main/java/com/thtf/environment/service/impl/AppEnvMonitorServiceImpl.java
@@ -66,7 +66,7 @@ public class AppEnvMonitorServiceImpl implements AppEnvMonitorService {
             tblAlarmRecordUnhandle.setBuildingCodeList(buildingCodesList);
             tblAlarmRecordUnhandle.setAlarmCategory(0);
             tblAlarmRecordUnhandle.setItemTypeCode(parameter.getItemTypeCode());
-            Integer data = alarmAPI.queryAllAlarmCount(tblAlarmRecordUnhandle).getData();
+            Long data = alarmAPI.queryAllAlarmCount(tblAlarmRecordUnhandle).getData();
             keyValueDTO.setValue(data);
             result.add(keyValueDTO);
         }
